requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

torch~=2.5.0.dev20240620
torchaudio~=2.4.0.dev20240620
numpy~=1.26.4
encodec~=0.1.1
pandas~=2.2.2
tqdm~=4.66.4
soundfile~=0.12.1
librosa~=0.10.2.post1
audiomentations~=0.36.0
PyYAML~=6.0.1
lightning~=2.3.0
torchmetrics~=1.4.0.post0